third_party/blink/web_tests/http/tests/security/sandbox-inherit-to-initial-document-expected.txt

CONSOLE WARNING: An iframe which has both allow-scripts and allow-same-origin for its sandbox attribute can escape its sandboxing.
Test how sandbox flags get inherited to initial document in subframe.

PASS
